app/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile app/requirements.in --output-file app/requirements.txt
blinker==1.9.0
    # via flask
click==8.2.0
    # via flask
flask==3.1.0
    # via -r app/requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
python-dotenv==1.1.0
    # via -r app/requirements.in
redis==6.0.0
    # via -r app/requirements.in
werkzeug==3.1.3
    # via flask
